requirements.txt
anyjson==0.3.3
blinker==1.3
celery==3.1.5
Django==1.6
django-celery==3.1.1
django-extensions==1.2.5
django-pagination==1.0.7
django-tastypie==0.10.0
elasticsearch==0.4.3
ipython==1.1.0
mimeparse==0.1.3
mongoengine==0.8.4
requests==2.0.1
South==0.8.4